# Improve performance of cached estimates in logs