## Description

This avoids instantiating two RepoStructure objects for each commit in a
log, when we're fetching cached estimates (stored in annotations.db) with `--with-feature-count=...`

This shaves a bit off processing each commit.
For a repo with 492 commits, the difference is ~0.43s:

Before:
```
$ hyperfine 'kart log --with-feature-count=medium -o json > /dev/null'
Benchmark #1: kart log --with-feature-count=medium -o json > /dev/null
  Time (mean ± σ):      2.861 s ±  0.034 s    [User: 2.078 s, System: 0.769 s]
  Range (min … max):    2.813 s …  2.919 s    10 runs
```

After:
```
$ hyperfine 'kart log --with-feature-count=medium -o json > /dev/null'
Benchmark #1: kart log --with-feature-count=medium -o json > /dev/null
  Time (mean ± σ):      2.429 s ±  0.043 s    [User: 1.851 s, System: 0.571 s]
  Range (min … max):    2.371 s …  2.494 s    10 runs
```

## Related links:

none

## Checklist:

- [x] Have you reviewed your own change?
- [ ] Have you included test(s)?
- [ ] Have you updated the [changelog](https://github.com/koordinates/kart/blob/master/CHANGELOG.md)?